Move date range util
