+++ crates/polars-time/src/windows/test.rs
@@ -2,6 +2,7 @@ use chrono::prelude::*;
 use polars_arrow::export::arrow::temporal_conversions::timestamp_ns_to_datetime;
 use polars_core::prelude::*;
 
+use crate::date_range::datetime_range_i64;
 use crate::prelude::*;
 
 #[test]
